AAUTO-1351 MX95:make bluetooth product.prop specific to non-Car builds
As the core/sysprop.mk contains wildcard, looking for product.prop file in $TARGET_DEVICE_DIR,
it includes the product.prop file even if we disable it in makefile. We need custom name for the
file thus the wildcard is not used by default.

This allows different bluetooth configuration for Car builds.

This is similar to AAUTO-750 which was discovered with imx8q

Change-Id: I9f0ea7f8a75ecaf4151488650ebcf07f8e8219d2
Signed-off-by: Ondrej Lutera <ondrej.lutera@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24557
Reviewed-by: Yunjie Jia <yunjie.jia@nxp.com>
Reviewed-by: Jiri Rezler <jiri.rezler@nxp.com>