MA-23352 optimze the flash image scripts related to card size op
the imx-sdcard-partition.sh cannot be used on evk_95 with the option
"-c 13" as 13 is not taken as a legal value by the script.

This patch makd 13 an legal value. also optimize the code to check
whether a card size is proper, both for imx-sdcard-partition.sh, and
other four flash scripts.

Change-Id: I5f828fc3d91037f55e9796efff4e2e90fa76a99f
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27359
Reviewed-by: Ji Luo <ji.luo@nxp.com>