MA-22740-2 erase metadata/userdata insteand of format
The make_f2fs options were dumped from fastboot command. If 16KB page
size is used, more options need to be appended. Also, it is found that
the system has better boot time performance if the partitions are
formatted during first time boot up. So just erase the metadata and
userdata partitions in this script.

Change-Id: Icec67d6a817fa02caaa7c2c7c0664242553f50db
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27361
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>