MA-23414 eliminate warning logs when build super.img
Previously, the android filesystem images like system.img are generated
as erofs filesystem image, not stored as android sparse image, then when
lpmake is used to generate super.img with these images, there are
following warning log print

  Invalid sparse file format at header magic

To eliminate the warning log, need to set either of the following values
to make the images stored as android sparse images.
1. partition size
2. partition reserved size
3. partition headroom

This commit choose to set the reserved size for partitions to be 0. as
in build_image.py, when read the reserved size of erofs partitions, if
it is not set, the value is default to be 0.

Change-Id: I94077509aa44c96c515bdef08686bc57fcdda718
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27509
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>