MA-23516 add back 'inlinecrypt' mount option for data
The 'inlinecrypt' mount option can be used to specify to
use inline encryption when possible, it was removed before
because there is no inline encryption hardware on i.MX.

However, the vellamo benchmark test application seems will
depend on this mount option and the storage I/O test will
fail without adding this mount option.

According to the fscrypt documentation in linux kernel, the
'inlinecrypt' will only specify to use inline encryption
when the support exists, it will fallback to use default
kernel crypto API or blk-crypto-fallback when the hardware
support is not available. So adding this mount option should
not cause any regression.

This commit adds the 'inlinecrypt' back to make the vellamo
happy.

Change-Id: Idee1e22f328809a138bc07ca683e25a4fbddbb22
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27671
Reviewed-by: Elven Wang <elven.wang@nxp.com>