MA-23956-1 fix selinux neverallow error with 25Q4 code
dumpstate is not allowed to call vold, so directly remove the rules to
allow vold to call vold.

for hal_secureime_service and hal_secure_enclave_service, they are not
associated with the attribute domain, and in 25Q4 code, a type that is
not associated with the attribute domain is not taken as a process, and
it is not allowed to access binder. In this commit, associate the two
types with attribute domain.

Change-Id: Ia6d11d2786e43ebb3124b8ffc80c38a3c26fa70e
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>