MA-24261 Include the dtb mapping to dtbs structure
When the dtbs are included to the vendor_boot image, they
are organized into dtb/dtbo structure[1]. We can specify
an "id" for each dtb but its "name" is too long to be stored
into the dtb/dtbo structure. However, we still need an "id"
to "name" mapping to make it easier to check/change the target
dtb.

It's tedious to maintain the mapping in both bootloader and
"device/nxp". This commit will design a new structure to hold
the "id" to "name" mapping, the resulting binary would be
packaged to the first dtb entry in the dtb/dtbo structure.
Bootloader should parse the mapping to get the target dtb id.

[1]: https://source.android.com/docs/core/architecture/dto/partitions

Change-Id: Id09de7cd895e1d60d855ba8964aa0e85a8e6d7df
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/30279
Reviewed-by: Maximus Sun <maximus.sun@nxp.com>